Attach MIT open source license
While there have been many forks of this project and it has been stated that it is open source, there does not appear to be any actual open source license attached to the code.  This means that contributors do not actually have a license to use the code they're contributing to.  This proposal is for the project to add a simple MIT open source license to the codebase.